+++ R/B_swFiles.R
@@ -57,7 +57,7 @@ setClass(
   prototype = list(
     ProjDir = NA_character_,
     # 29 must be equal to rSW2_glovars[["kSOILWAT2"]][["kINT"]][["SW_NFILES"]]
-    InFiles = rep(NA_character_, 29L),
+    InFiles = rep(NA_character_, 31L),
     WeatherPrefix = NA_character_,
     OutputPrefix = NA_character_
   )
@@ -141,7 +141,7 @@ setMethod(
     is_lt_v600 <- n_has == 22L
     is_in_v600tov604 <- n_has == 23L
     is_ge_v610 <- n_has == 27L
-    is_ge_v670 <- n_has == 29L
+    is_ge_v670 <- n_has == 31L
 
     #--- Upgrade object to v6.0.0-v6.0.4: from < v6.0.0
     if (is_lt_v600 && n_exp == 23L) {
@@ -156,6 +156,7 @@ setMethod(
     if (is_lt_v600 && n_exp == 27L) {
       object@InFiles <- c(
         object@InFiles[1L],
+        # object@InFiles[2L] "years.in" was dropped with v6.1.0
         target@InFiles[2L:6L], # insert names of new files with v6.1.0
         object@InFiles[3L:5L],
         target@InFiles[10L], # insert name of new file with v6.0.0,
@@ -167,41 +168,60 @@ setMethod(
     if (is_in_v600tov604 && n_exp == 27L) {
       object@InFiles <- c(
         object@InFiles[1L],
+        # object@InFiles[2L] "years.in" was dropped with v6.1.0
         target@InFiles[2L:6L], # insert names of new files with v6.1.0
         object@InFiles[3L:23L]
       )
     }
 
     #--- Upgrade object to >= v6.7.0: from < v6.0.0
-    if (is_lt_v600 && n_exp == 29L) {
+    if (is_lt_v600 && n_exp == 31L) {
       object@InFiles <- c(
         object@InFiles[1L],
+        # object@InFiles[2L] "years.in" was dropped with v6.1.0
         target@InFiles[2L:4L], # insert names of new files with v6.1.0
         target@InFiles[5L:6L], # insert names of new files with v6.7.0
         target@InFiles[7L:8L], # insert names of new files with v6.1.0
         object@InFiles[3L:5L],
         target@InFiles[12L], # insert name of new file with v6.0.0
-        object@InFiles[6L:22L]
+        object@InFiles[6L:17L],
+        target@InFiles[25L], # insert new "Output/sw2_seasonal.csv" with v6.7.0
+        object@InFiles[18L:21L],
+        # insert new "Output/sw2_seasonal_slyrs.csv" with v6.7.0
+        target@InFiles[30L],
+        object@InFiles[22L]
       )
     }
 
     #--- Upgrade object to >= v6.7.0: from v6.0.0-v6.0.4
-    if (is_in_v600tov604 && n_exp == 29L) {
+    if (is_in_v600tov604 && n_exp == 31L) {
       object@InFiles <- c(
         object@InFiles[1L],
+        # object@InFiles[2L] "years.in" was dropped with v6.1.0
         target@InFiles[2L:4L], # insert names of new files with v6.1.0
         target@InFiles[5L:6L], # insert names of new files with v6.7.0
         target@InFiles[7L:8L], # insert names of new files with v6.1.0
-        object@InFiles[3L:23L]
+        object@InFiles[3L:18L],
+        target@InFiles[25L], # insert new "Output/sw2_seasonal.csv" with v6.7.0
+        object@InFiles[19L:22L],
+        # insert new "Output/sw2_seasonal_slyrs.csv" with v6.7.0
+        target@InFiles[30L],
+        object@InFiles[23L]
       )
     }
 
     #--- Upgrade object to >= v6.7.0: from v6.1.0-v6.6.x
-    if (is_ge_v610 && n_exp == 29L) {
+    if (is_ge_v610 && n_exp == 31L) {
       object@InFiles <- c(
         object@InFiles[1L:4L],
-        target@InFiles[5L:6L], # insert names of new files with v6.7.0
-        object@InFiles[5L:27L]
+        # insert "Input_nc/cached_state.nc" and "Input_nc/system.in" with v6.7.0
+        target@InFiles[5L:6L],
+        object@InFiles[5L:22L],
+        target@InFiles[25L], # insert new "Output/sw2_seasonal.csv" with v6.7.0
+        object@InFiles[23L:26L],
+        # insert new "Output/sw2_seasonal_slyrs.csv" with v6.7.0
+        target@InFiles[30L],
+        object@InFiles[27L]
       )
     }